+++ src/factorii/ui/content.h
@@ -6,7 +6,7 @@
 #include <factorii/ui/helpers.h>
 #include <factorii/ui/nuk.h>
 
-void fii_content(GLFWwindow *win, struct nk_context *ctx, int width,
-                 int height);
+void fii_content(GLFWwindow *win, struct nk_context *ctx, float width,
+                 float height);
 
 #endif
