+++ src/factorii/ui/fonts.h
@@ -4,14 +4,14 @@
 #include <stdio.h>
 #include <ubelt/utils.h>
 
-#include <ft2build.h>
-#include <freetype/freetype.h>
-#include <freetype/ftglyph.h>
-#include <freetype/ftoutln.h>
-#include <freetype/fttrigon.h>
+//#include <ft2build.h>
+//#include <freetype/freetype.h>
+//#include <freetype/ftglyph.h>
+//#include <freetype/ftoutln.h>
+//#include <freetype/fttrigon.h>
 
 #include <factorii/build.h>
 
 void fii_fonts_load(struct nk_context *ctx);
-struct nk_user_font fii_font(struct nk_context *ctx, int size);
+struct nk_user_font fii_font(struct nk_context *ctx, float size);
 #endif
